refactor(eval): use shared IO utilities
