+++ src/handlers/eval/evaluator/llm-as-a-judge/index.tsx
@@ -1,5 +1,6 @@
 import { Router } from "../../../../router";
-import type { AppIO, Core } from "../../../types";
+import type { AppIO } from "../../../../io";
+import type { Core } from "../../../types";
 import { createHelpDefault } from "../../../help";
 import { createLlmAsAJudgeCreateHandler } from "./create";
 import { createLlmAsAJudgeUpdateHandler } from "./update";
