+++ src/handlers/eval/index.tsx
@@ -1,5 +1,6 @@
 import { Router } from "../../router";
-import type { AppIO, Core } from "../types";
+import type { AppIO } from "../../io";
+import type { Core } from "../types";
 import { createHelpDefault } from "../help";
 import { createEvaluatorHandler } from "./evaluator";
